# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: goimg-datalayer API
  version: 2.0.0
  description: |
    A self-hosted image gallery backend inspired by Flickr and Chevereto.

    Features:
    - User management with JWT authentication
    - Image upload with automatic variant generation
    - Albums, tags, and social features (likes, comments)
    - Content moderation and abuse reporting
    - RESTful API with RFC 7807 error responses

    **Phase 2 Complete**: All Phase 2 features (Sprints 10-15) implemented.
    **Phase 3 Progress**: Sprints 16-20 complete (oEmbed, nested albums, custom variants, test coverage, trending tags, featured picks, groups/communities).
    **Current Sprint**: Sprint 23 in progress (test coverage improvement, documentation updates).
    Future features in backlog: video support (Sprint 21) and account tiers (Sprint 22).

  contact:
    name: goimg-datalayer API Team
    url: https://github.com/yegamble/goimg-datalayer
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://api.goimg.com/v1
    description: Production server (example)
security:
  - bearerAuth: []
...
paths:
...
  /users/{id}/likes:
    get:
      tags: [users]
      summary: Get user's liked images
      description: |
        Retrieve a paginated list of images that the specified user has liked.

        Access control:
        - Public users: Can view likes of users with public profiles
        - Authenticated users: Can view their own likes and public likes
      operationId: getUserLikedImages
      security:
        - {}
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
      responses:
        '200':
          description: Liked images retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/Image'
        '401':
          description: Unauthorized (optional - only if authentication required)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

  # ==================== User Follow/Unfollow ====================
...
  /images/{id}/like:
    post:
      tags: [social]
      summary: Like image
      description: Like an image. This operation is idempotent.
      operationId: likeImage
      parameters:
        - $ref: '#/components/parameters/ImageIdParam'
      responses:
        '200':
          description: Image liked successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  liked:
                    type: boolean
                    example: true
                  like_count:
                    type: integer
                    example: 42
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '404':
          description: Image not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

    delete:
      tags: [social]
      summary: Unlike image
      description: Remove like from an image
      operationId: unlikeImage
      parameters:
        - $ref: '#/components/parameters/ImageIdParam'
      responses:
        '200':
          description: Image unliked successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  liked:
                    type: boolean
                    example: false
                  like_count:
                    type: integer
                    example: 41
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '404':
          description: Image not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

  # NOTE: GET /images/{id}/likes endpoint planned for future release
  # Will list users who liked an image with pagination
  # Requires: LikeRepository.GetUsersWhoLiked() method implementation

  # ==================== Social (Comments) ====================
  /images/{id}/comments:
    post:
      tags: [social]
      summary: Add comment
      description: Add a comment to an image
      operationId: addComment
      parameters:
        - $ref: '#/components/parameters/ImageIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
                  minLength: 1
                  maxLength: 1000
                  example: Beautiful photo! Love the colors.
      responses:
        '201':
          description: Comment added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '404':
          description: Image not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

    get:
      tags: [social]
      summary: List comments
      description: Retrieve all comments for an image
      operationId: listComments
      security:
        - {}
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ImageIdParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
      responses:
        '200':
          description: Comments retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/Comment'
  /comments/{id}:
    delete:
      tags: [social]
      summary: Delete comment
      description: |
        Delete a comment. Can be deleted by:
        - Comment author
        - Image owner
        - Moderator/Admin
      operationId: deleteComment
      parameters:
        - name: id
          in: path
          required: true
          description: Comment ID
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Comment deleted successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '403':
          description: Forbidden - cannot delete this comment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '404':
          description: Comment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

  # ==================== Moderation ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer authentication. Include the access token in the Authorization header:

        ```
        Authorization: Bearer <access_token>
        ```

        Access tokens expire after 15 minutes. Use the refresh endpoint to obtain a new token.
  parameters:
    UserIdParam:
      name: id
      in: path
      required: true
      description: User ID
      schema:
        type: string
        format: uuid
      example: 123e4567-e89b-12d3-a456-426614174000
    ImageIdParam:
      name: id
      in: path
      required: true
      description: Image ID
      schema:
        type: string
        format: uuid
      example: 123e4567-e89b-12d3-a456-426614174000
...
    PageParam:
      name: page
      in: query
      description: Page number
      schema:
        type: integer
        minimum: 1
        default: 1
      example: 1
    PerPageParam:
      name: per_page
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      example: 20
...
  schemas:
...
    Image:
      type: object
      required: [id, owner_id, owner, visibility, mime_type, width, height, variants, created_at]
      properties:
        id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        owner_id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174001
        owner:
          type: object
          required: [id, username]
          properties:
            id:
              type: string
              format: uuid
            username:
              type: string
            display_name:
              type: string
              nullable: true
          example:
            id: 123e4567-e89b-12d3-a456-426614174001
            username: johndoe
            display_name: John Doe
        title:
          type: string
          nullable: true
          example: Sunset over mountains
        description:
          type: string
          nullable: true
          example: Beautiful sunset captured in Rocky Mountains
        visibility:
          type: string
          enum: [public, private, unlisted]
          example: public
        mime_type:
          type: string
          example: image/jpeg
        file_size:
          type: integer
          description: File size in bytes
          example: 2048576
        width:
          type: integer
          description: Image width in pixels
          example: 4000
        height:
          type: integer
          description: Image height in pixels
          example: 3000
        variants:
          type: object
          properties:
            thumbnail:
              $ref: '#/components/schemas/ImageVariant'
            small:
              $ref: '#/components/schemas/ImageVariant'
            medium:
              $ref: '#/components/schemas/ImageVariant'
            large:
              $ref: '#/components/schemas/ImageVariant'
            original:
              $ref: '#/components/schemas/ImageVariant'
        tags:
          type: array
          items:
            type: string
          example: [sunset, mountains, nature, landscape]
        view_count:
          type: integer
          example: 150
        like_count:
          type: integer
          example: 42
        comment_count:
          type: integer
          example: 8
        created_at:
          type: string
          format: date-time
          example: 2024-01-15T10:30:00Z
        updated_at:
          type: string
          format: date-time
          example: 2024-01-15T11:00:00Z
    ImageVariant:
      type: object
      required: [url, width, height]
      properties:
        url:
          type: string
          format: uri
          example: https://cdn.goimg.com/images/123e4567/medium.jpg
        width:
          type: integer
          example: 800
        height:
          type: integer
          example: 600

    # ==================== IPFS Schemas ====================
...
    Comment:
      type: object
      required: [id, user_id, user, image_id, content, created_at]
      properties:
        id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        user_id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174001
        user:
          type: object
          required: [id, username]
          properties:
            id:
              type: string
              format: uuid
            username:
              type: string
            display_name:
              type: string
              nullable: true
          example:
            id: 123e4567-e89b-12d3-a456-426614174001
            username: johndoe
            display_name: John Doe
        image_id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174002
        content:
          type: string
          example: Beautiful photo! Love the colors.
        created_at:
          type: string
          format: date-time
          example: 2024-01-15T10:30:00Z
...
    PaginatedResponse:
      type: object
      required: [items, pagination]
      properties:
        items:
          type: array
          items: {}
        pagination:
          $ref: '#/components/schemas/Pagination'
    Pagination:
      type: object
      required: [total, page, per_page, total_pages]
      properties:
        total:
          type: integer
          description: Total number of items
          example: 150
        page:
          type: integer
          description: Current page number
          example: 1
        per_page:
          type: integer
          description: Items per page
          example: 20
        total_pages:
          type: integer
          description: Total number of pages
          example: 8

    # ==================== Error Responses (RFC 7807) ====================
    ProblemDetail:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          description: A URI reference that identifies the problem type
          example: https://api.goimg.com/problems/validation-error
        title:
          type: string
          description: A short, human-readable summary of the problem type
          example: Validation Error
        status:
          type: integer
          description: The HTTP status code
          example: 400
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence
          example: Request validation failed
        instance:
          type: string
          format: uri
          description: A URI reference that identifies the specific occurrence
          example: /api/v1/images
        traceId:
          type: string
          description: A unique identifier for tracing the request
          example: abc123-def456
        errors:
          type: array
          description: Detailed validation errors (for 400 responses)
          items:
            type: object
            properties:
              field:
                type: string
                example: title
              message:
                type: string
                example: must not exceed 255 characters

    # ==================== Session ====================
...